DONE : checkboxes changing state and the filelisting now changing row color to red if unchecked. preview button now shows that clicked dataset in preview datagrid. within the datagrid for the results, the thresholds are taken from the readings and colored to show low or high thresholds crossed.
TODO :
need to add a comments box, so that it can be added to the test results object before putting to PDF.
need to have a few boxes to show and change threshold values.

WiSHLIST : possibly add a color swatch down the line if they want to change the threshold colors.